Stop and mask the ntpd instead of remove it
In some cases when the NTP is dependent on other packages we have
a situation that we have to remove these packages in order to remove
the NTP which is not the main idea of this project.
thus IMHO it will be better to stop and mask the NTP instead of
remove it.

Note:
The block feature introduced in ansible 2.0
The systemd module introduced in ansible 2.2